fix gis_tools.py utm_zone data type issue for py3 unit test and bins must be an integer in plotstrike.py

+++ mtpy/imaging/plotstrike.py
@@ -478,7 +478,7 @@ class PlotStrike(object):
                     tr = medtipr[binlist, :]
                     # compute the historgram for the tipper strike
                     trhist = np.histogram(tr[np.nonzero(tr)].flatten(),
-                                          bins=360 / bw,
+                                          bins=int(360/bw),
                                           range=histrange)
 
                     # make a bar graph with each bar being width of bw degrees
@@ -497,10 +497,10 @@ class PlotStrike(object):
 
                 # estimate the histogram for the decade for invariants and pt
                 invhist = np.histogram(hh[np.nonzero(hh)].flatten(),
-                                       bins=360 / bw,
+                                       bins= int(360/bw),
                                        range=histrange)
                 pthist = np.histogram(ptplotdata,
-                                      bins=360 / bw,
+                                      bins=int(360/bw),
                                       range=histrange)
         
                 # plot the histograms
@@ -730,7 +730,7 @@ class PlotStrike(object):
 
             # estimate the histogram for the decade for invariants and pt
             invhist = np.histogram(hh[np.nonzero(hh)].flatten(),
-                                   bins=360 / bw,
+                                   bins=int(360/bw),
                                    range=histrange)
 
             ptplotdata = gg[np.nonzero(gg)].flatten()
@@ -755,7 +755,7 @@ class PlotStrike(object):
                 
 
             pthist = np.histogram(ptplotdata,
-                                  bins=360 / bw,
+                                  bins=int(360/bw),
                                   range=histrange
                                   )
 
@@ -783,7 +783,7 @@ class PlotStrike(object):
                 tr = self._medtp[binlist, :]
 
                 trhist = np.histogram(tr[np.nonzero(tr)].flatten(),
-                                      bins=360 / bw,
+                                      bins= int(360/bw),
                                       range=histrange
                                       )
 
@@ -1131,7 +1131,7 @@ class PlotStrike(object):
                     pass
 
                 zshist = np.histogram(zs[np.nonzero(zs)].flatten(),
-                                      bins=360 / bw,
+                                      bins=int(360/bw),
                                       range=histrange)
 
                 #==============================================================
@@ -1186,7 +1186,7 @@ class PlotStrike(object):
 
                 # == > compute mode
                 azhist = np.histogram(az[np.nonzero(az)].flatten(),
-                                      bins=360 / bw,
+                                      bins=int(360/bw),
                                       range=histrange)
                 ptmode1 = 90 - azhist[1][np.where(
                     azhist[0] == azhist[0].max())[0][0]]
@@ -1217,7 +1217,7 @@ class PlotStrike(object):
                     tipr[np.where(tipr < 0)] = tipr[np.where(tipr < 0)] + 360
 
                 tphist = np.histogram(tipr[np.nonzero(tipr)].flatten(),
-                                      bins=360 / bw,
+                                      bins=int(360/bw),
                                       range=histrange)
 
                 #==> compute mean
@@ -1256,14 +1256,14 @@ class PlotStrike(object):
 
             # estimate the histogram for the decade for invariants and pt
             invhist = np.histogram(hh[np.nonzero(hh)].flatten(),
-                                   bins=360 / bw,
+                                   bins=int(360/bw),
                                    range=histrange)
             pthist = np.histogram(ptplotdata,
-                                  bins=360 / bw,
+                                  bins=int(360/bw),
                                   range=histrange)
 
             trhist = np.histogram(tr[np.nonzero(tr)].flatten(),
-                                  bins=360 / bw,
+                                  bins=int(360/bw),
                                   range=histrange)
 
             #--> include the row for mean, median and mode for each parameter
